% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.mcmc.list.bvsOccupancy}
\alias{as.mcmc.list.bvsOccupancy}
\title{Posterior samples as a coda object}
\usage{
\method{as.mcmc.list}{bvsOccupancy}(x, ...)
}
\arguments{
\item{x}{A fitted model from \code{\link{bvsSSOM}}, \code{\link{bvsSDOM}} or
\code{\link{bvsMSOM}}.}

\item{...}{Ignored.}
}
\value{
An \code{\link[coda]{mcmc.list}}, ready for \pkg{coda} or
  \pkg{MCMCvis}.
}
\description{
Posterior samples as a coda object
}
