% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble-polyagamma.R
\name{aterm}
\alias{aterm}
\title{Coefficient of the alternating series used in the Polya-Gamma sampler}
\usage{
aterm(n = integer(), x = double(0), t = double(0))
}
\arguments{
\item{n, x, t, mu, z}{Arguments of the underlying algorithm.}
}
\value{
A scalar draw.
}
\description{
Internal building block of the exact Polya-Gamma sampler of Polson, Scott
and Windle (2013); see \code{\link{samplepg}}.
}
\keyword{internal}
