% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers-colonisation.R
\name{gamma_sampler_eta}
\alias{gamma_sampler_eta}
\title{Variable selection sampler for colonisation}
\usage{
gamma_sampler_eta(model, mvSaved, target, control)
}
\arguments{
\item{model}{The model object.}

\item{mvSaved}{A \code{modelValues} object used to restore the model state.}

\item{target}{The node being sampled.}

\item{control}{A named list of control parameters:
\describe{
\item{ncov}{Number of covariates excluding the intercept.}
\item{indexes_covariates}{Integer vector mapping columns of \code{designMatrix} to covariate groups.}
\item{designMatrix}{Design matrix, with rows aligned to the nodes named below.}
\item{PG}{Name of the Polya-Gamma node.}
\item{fixedEffects}{Name of the regression coefficient node.}
\item{znodes}{Name of the latent occupancy node.}
\item{ynodes}{Name of the occupancy node one season ahead. Optional; when omitted the stochastic dependencies of \code{target} are used instead.}
}}
}
\value{
A \pkg{nimble} sampler object.
}
\description{
Add / delete / swap Metropolis-Hastings update of the inclusion vector for the colonisation component of a dynamic occupancy model. Only sites that were unoccupied in the current season contribute.
}
\section{Use}{

This sampler is assigned automatically by \code{\link{bvsSSOM}},
\code{\link{bvsSDOM}} and \code{\link{bvsMSOM}}. It is exported so that
it can be added by hand to a custom \pkg{nimble} model configuration with
\code{conf$addSampler()}.
}

