% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble-helpers.R
\name{mvrnormQuick}
\alias{mvrnormQuick}
\title{Fast multivariate normal draw from a Cholesky factor}
\usage{
mvrnormQuick(mu = double(1), cholsigma = double(2))
}
\arguments{
\item{mu}{Mean vector.}

\item{cholsigma}{Cholesky factor of the covariance matrix.}
}
\value{
A single draw as a numeric vector.
}
\description{
Fast multivariate normal draw from a Cholesky factor
}
\keyword{internal}
