% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble-polyagamma.R
\name{pgr}
\alias{pgr}
\title{Draw from a Polya-Gamma PG(n, z) distribution}
\usage{
pgr(n = integer(), z = double(0))
}
\arguments{
\item{n}{Integer shape parameter, the number of Bernoulli trials.}

\item{z}{Tilting parameter.}
}
\value{
A single Polya-Gamma draw.
}
\description{
Sums \code{n} independent PG(1, z) draws.
}
\keyword{internal}
