% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble-polyagamma.R
\name{randinvg}
\alias{randinvg}
\title{Draw from an inverse Gaussian distribution}
\usage{
randinvg(mu = double(0))
}
\arguments{
\item{n, x, t, mu, z}{Arguments of the underlying algorithm.}
}
\value{
A scalar draw.
}
\description{
Internal building block of the exact Polya-Gamma sampler of Polson, Scott
and Windle (2013); see \code{\link{samplepg}}.
}
\keyword{internal}
