% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble-polyagamma.R
\name{samplepg}
\alias{samplepg}
\title{Exact draw from a Polya-Gamma PG(1, z) distribution}
\usage{
samplepg(z = double(0))
}
\arguments{
\item{z}{Tilting parameter, typically the current linear predictor.}
}
\value{
A single Polya-Gamma draw.
}
\description{
Implements the accept-reject sampler of Polson, Scott and Windle (2013),
following the algorithm on page 130 of Windle's thesis.
}
\references{
Polson, N. G., Scott, J. G. and Windle, J. (2013) Bayesian inference for
logistic models using Polya-Gamma latent variables.
\emph{Journal of the American Statistical Association} \strong{108}, 1339--1349.
}
\keyword{internal}
